+++ app/controllers/observation_matrix_row_items_controller.rb
@@ -113,8 +113,7 @@ class ObservationMatrixRowItemsController < ApplicationController
   def observation_matrix_row_item_params
     params.require(:observation_matrix_row_item).permit(
       :observation_matrix_id, :type,
-      :collection_object_id, :otu_id,
-      :taxon_name_id,
-      :controlled_vocabulary_term_id, :type, :position)
+      :observation_object_type, :observation_object_id,
+      :type, :position)
   end
 end