+++ app/javascript/vue/tasks/observation_matrices/matrix_row_coder/store/mutations/setContinuousValue.js
@@ -4,9 +4,13 @@ import setDescriptorUnsaved from '../helpers/setDescriptorUnsaved'
 export default function (state, args) {
   const {
     descriptorId,
-    continuousValue
+    continuousValue,
+    observationId
   } = args
 
-  mergeIntoObservation(state.observations.find(o => o.descriptorId === descriptorId), { continuousValue })
+  mergeIntoObservation(state.observations.find(o => o.descriptorId === descriptorId && (o.id === observationId || o.internalId === observationId)), { 
+    continuousValue,
+    isUnsaved: true
+  })
   setDescriptorUnsaved(state.descriptors.find(d => d.id === descriptorId))
-};
+}