+++ app/models/concerns/shared/is_data.rb
@@ -28,6 +28,10 @@ module Shared::IsData
       self < Shared::IsDwcOccurrence
     end
 
+    def is_observable?
+      self < Shared::Observations
+    end
+
     # @return [Array] of strings of only the non-cached and non-housekeeping column names
     def data_attributes
       column_names.reject { |c| %w{id project_id created_by_id updated_by_id created_at updated_at}