+++ app/models/controlled_vocabulary_term.rb
@@ -48,12 +48,10 @@ class ControlledVocabularyTerm < ApplicationRecord
 
   validate :uri_relation_is_a_skos_relation, unless: -> {uri_relation.blank?}
 
-  has_many :observation_matrix_row_items, inverse_of: :controlled_vocabulary_term, class_name: 'ObservationMatrixRowItem::Dynamic::Tag', dependent: :destroy
+  has_many :observation_matrix_row_items, as: :observation_object, inverse_of: :observation_object,  class_name: 'ObservationMatrixRowItem::Dynamic::Tag', dependent: :destroy
   has_many :observation_matrix_column_items, inverse_of: :controlled_vocabulary_term, class_name: 'ObservationMatrixColumnItem::Dynamic::Tag', dependent: :destroy
   
   has_many :observation_matrices, through: :observation_matrix_row_items
-
-
 
   scope :of_type, -> (type) { where(type: type.to_s.capitalize) } # TODO, capitalize is not the right method for things like `:foo_bar`