+++ app/models/otu.rb
@@ -75,13 +75,6 @@ class Otu < ApplicationRecord
 
   has_many :content_topics, through: :contents, source: :topic
 
-  has_many :observation_matrix_row_items, inverse_of: :otu, dependent: :delete_all, class_name: 'ObservationMatrixRowItem::Single::Otu'
-  has_many :observation_matrix_rows, inverse_of: :collection_object, dependent: :delete_all
-  has_many :observation_matrices, through: :observation_matrix_rows
-
-  has_many :observations, inverse_of: :otu, dependent: :restrict_with_error
-  has_many :descriptors, through: :observations
-
   scope :with_taxon_name_id, -> (taxon_name_id) { where(taxon_name_id: taxon_name_id) }
   scope :with_name, -> (name) { where(name: name) }